# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
      "PerpetualPositionStatus": {
        "enum": [
          "OPEN",
          "CLOSED",
          "LIQUIDATED"
        ],
        "type": "string"
      },
      "PositionSide": {
        "enum": [
          "LONG",
          "SHORT"
        ],
        "type": "string"
      },
      "IsoString": {
        "type": "string"
      },
      "PerpetualPositionResponseObject": {
        "properties": {
          "market": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PerpetualPositionStatus"
          },
          "side": {
            "$ref": "#/components/schemas/PositionSide"
          },
          "size": {
            "type": "string"
          },
          "maxSize": {
            "type": "string"
          },
          "entryPrice": {
            "type": "string"
          },
          "realizedPnl": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "createdAtHeight": {
            "type": "string"
          },
          "sumOpen": {
            "type": "string"
          },
          "sumClose": {
            "type": "string"
          },
          "netFunding": {
            "type": "string"
          },
          "unrealizedPnl": {
            "type": "string"
          },
          "perpYieldIndex": {
            "type": "string"
          },
          "closedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/IsoString"
              }
            ],
            "nullable": true
          },
          "exitPrice": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "market",
          "status",
          "side",
          "size",
          "maxSize",
          "entryPrice",
          "realizedPnl",
          "createdAt",
          "createdAtHeight",
          "sumOpen",
          "sumClose",
          "netFunding",
          "unrealizedPnl",
          "perpYieldIndex"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "PerpetualPositionsMap": {
        "properties": {},
        "additionalProperties": {
          "$ref": "#/components/schemas/PerpetualPositionResponseObject"
        },
        "type": "object"
      },
      "AssetPositionResponseObject": {
        "properties": {
          "symbol": {
            "type": "string"
          },
          "side": {
            "$ref": "#/components/schemas/PositionSide"
          },
          "size": {
            "type": "string"
          },
          "assetId": {
            "type": "string"
          },
          "subaccountNumber": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "symbol",
          "side",
          "size",
          "assetId",
          "subaccountNumber"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "AssetPositionsMap": {
        "properties": {},
        "additionalProperties": {
          "$ref": "#/components/schemas/AssetPositionResponseObject"
        },
        "type": "object"
      },
      "SubaccountResponseObject": {
        "properties": {
          "address": {
            "type": "string"
          },
          "subaccountNumber": {
            "type": "number",
            "format": "double"
          },
          "equity": {
            "type": "string"
          },
          "freeCollateral": {
            "type": "string"
          },
          "openPerpetualPositions": {
            "$ref": "#/components/schemas/PerpetualPositionsMap"
          },
          "assetPositions": {
            "$ref": "#/components/schemas/AssetPositionsMap"
          },
          "marginEnabled": {
            "type": "boolean"
          },
          "assetYieldIndex": {
            "type": "string"
          }
        },
        "required": [
          "address",
          "subaccountNumber",
          "equity",
          "freeCollateral",
          "openPerpetualPositions",
          "assetPositions",
          "marginEnabled",
          "assetYieldIndex"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "AddressResponse": {
        "properties": {
          "subaccounts": {
            "items": {
              "$ref": "#/components/schemas/SubaccountResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "subaccounts"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "ParentSubaccountResponse": {
        "properties": {
          "address": {
            "type": "string"
          },
          "parentSubaccountNumber": {
            "type": "number",
            "format": "double"
          },
          "equity": {
            "type": "string"
          },
          "freeCollateral": {
            "type": "string"
          },
          "childSubaccounts": {
            "items": {
              "$ref": "#/components/schemas/SubaccountResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "address",
          "parentSubaccountNumber",
          "equity",
          "freeCollateral",
          "childSubaccounts"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Klyra Indexer API",
    "version": "v1.0.0",
    "description": "Scroll down for code samples, example requests and responses.\nBase URLs:\n* For **Testnet**, use <a href=\"https://klyra-testnet.imperator.co/v4\">https://klyra-testnet.imperator.co/v4</a>\nNote: Messages on Indexer WebSocket feeds are typically more recent than data fetched via Indexer's REST API, because the latter is backed by read replicas of the databases that feed the former. Ordinarily this difference is minimal (less than a second), but it might become prolonged under load."
  },
  "openapi": "3.0.0",
  "paths": {
    "/addresses/{address}": {
      "get": {
        "operationId": "GetAddress",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/addresses/{address}/subaccountNumber/{subaccountNumber}": {
      "get": {
        "operationId": "GetSubaccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubaccountResponseObject"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "subaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/addresses/{address}/parentSubaccountNumber/{parentSubaccountNumber}": {
      "get": {
        "operationId": "GetParentSubaccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParentSubaccountResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "parentSubaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://klyra-testnet.imperator.co/v4",
      "description": "Testnet"
    }
  ]
}